ensure first run is scoped soley to first run
Object.create is the proper way to create a new object quickly.
http://stackoverflow.com/posts/9772788/revisions

This breaks the reference counting that was causing a bug.
By default in js objects are passed by reference, which meant that all
requests were using the firstFetchCount number instead of specified
maxCount.

This change forces maxCount to be used as expected after
firstFetchCount is used only in the first run.