handle login failures better
This should help Dexcom's servers.
They should not be issuing 500, the correct response would be 401.

This patch fixes several bugs in handling failures and error handling from
Dexcom's servers.
In particular, this patch fixes the test to determine the success state of the
login attempt.  Previously, the script would frenetically try to login,
ignoring the sleep time until the next tick.  Fixing the test to only run again
in the case of success causes the script to properly wait until the subsequent
tick interval before resuming.

Secondly, this change now causes the script to abort all operations if the
number of failures exceeds the `maxFailures` environment variable.