make reentrant, safe for multi-tenant-use
This patch carries the accountId in the accumulator/opts, rather than using a
global.  This should make it reentrant and safe for multi-tenant-use.
Currently, when authorize is called repeatedly with different options, it will
cause locked accounts, and automated password resets due to use of the module
level global variable for tracking the accountId.  This change eliminate the
global, which should make it safe for callers to call authorize with different
account credentials.